Update Prior and Posterior Weights Given \(\pi_0\) and \(\pi_{alt}\)

This function updates the prior and posterior weights in a FASH model using the estimated proportion of null datasets (\(\pi_0\)) and the reweighted prior under the alternative hypothesis (\(\pi_{alt}\)).

fash_prior_posterior_update(L_matrix, pi0, pi_alt, grid)

Arguments

L_matrix

A numeric matrix representing the log-likelihoods of datasets across mixture components. Rows correspond to datasets, and columns correspond to mixture components.

pi0

A numeric scalar representing the estimated proportion of null datasets.

pi_alt

A numeric vector representing the estimated weights of the alternative components.

grid

A numeric vector representing the grid of Predictive Standard Deviation (PSD) values.

Value

A list containing:

prior_weight

A data frame with two columns:

psd

A numeric vector of PSD values corresponding to non-trivial weights.

prior_weight

A numeric vector of prior weights corresponding to the PSD values.

posterior_weight

A numeric matrix of posterior weights, where rows correspond to datasets and columns correspond to non-trivial mixture components.

Examples

# Example usage:
set.seed(1)
L_matrix <- matrix(rnorm(50), nrow = 10, ncol = 5)
pi0_hat <- 0.8
pi_alt <- rep(0.2, 4)  # Alternative weights
grid <- seq(0, 2, length.out = 5)
update_result <- fashr:::fash_prior_posterior_update(L_matrix, pi0_hat, pi_alt, grid)

# View updated prior weights
print(update_result$prior_weight)
#>   psd prior_weight
#> 1 0.0         0.80
#> 2 0.5         0.04
#> 3 1.0         0.04
#> 4 1.5         0.04
#> 5 2.0         0.04

# View updated posterior weights
print(update_result$posterior_weight)
#>               0         0.5           1         1.5          2
#>  [1,] 0.4757193 0.201810828 0.111556029 0.173162273 0.03775153
#>  [2,] 0.8181538 0.050275506 0.074426426 0.030719175 0.02642507
#>  [3,] 0.6298726 0.039023861 0.078255480 0.107027465 0.14582062
#>  [4,] 0.9710593 0.001075381 0.001347192 0.009333120 0.01718501
#>  [5,] 0.8300498 0.091943712 0.055482707 0.007532192 0.01499160
#>  [6,] 0.7424196 0.080617608 0.079720116 0.055681955 0.04156072
#>  [7,] 0.8917264 0.026945270 0.023434342 0.018461881 0.03943209
#>  [8,] 0.8764620 0.053821476 0.004811861 0.019737551 0.04516712
#>  [9,] 0.8396854 0.053663594 0.014634234 0.070919043 0.02109777
#> [10,] 0.6513401 0.080045036 0.067129908 0.094808954 0.10667596