Constructing and evaluating the local O-spline basis (design matrix)

Constructing and evaluating the local O-spline basis (design matrix)

local_poly_helper(knots, refined_x, p = 2, neg_sign_order = 0)

Arguments

knots

A vector of knots used to construct the O-spline basis, first knot should be viewed as "0", the reference starting location. These k knots will define (k-1) basis function in total.

refined_x

A vector of locations to evaluate the O-spline basis

p

An integer value indicates the order of smoothness

neg_sign_order

An integer value N such that D = ((-1)^N)*D for the splines at negative knots.

Value

A matrix with i,j component being the value of jth basis function value at ith element of refined_x, the ncol should equal to number of knots minus 1, and nrow should equal to the number of elements in refined_x.

Examples

fashr:::local_poly_helper(knots = c(0, 0.2, 0.4, 0.6, 0.8), refined_x = seq(0, 0.8, by = 0.1), p = 2)
#>        [,1]  [,2]  [,3]         [,4]
#>  [1,] 0.000 0.000 0.000 0.000000e+00
#>  [2,] 0.005 0.000 0.000 0.000000e+00
#>  [3,] 0.020 0.000 0.000 0.000000e+00
#>  [4,] 0.040 0.005 0.000 0.000000e+00
#>  [5,] 0.060 0.020 0.000 0.000000e+00
#>  [6,] 0.080 0.040 0.005 0.000000e+00
#>  [7,] 0.100 0.060 0.020 6.162976e-33
#>  [8,] 0.120 0.080 0.040 5.000000e-03
#>  [9,] 0.140 0.100 0.060 2.000000e-02