Simulate Observed Data from a Specified True Function

This function generates synthetic observed data by evaluating a user-specified true function g on a grid of x-values and adding Gaussian noise with heteroscedastic standard deviation sampled from a provided vector or scalar.

simulate_data(g, x = NULL, sd = 0.1)

Arguments

g

A function specifying the true underlying signal to simulate.

x

A numeric vector of grid points where the function g is evaluated.

sd

A numeric value or vector specifying the standard deviation(s) of noise. If a scalar, all points have the same standard deviation. If a vector, it is sampled with replacement for each x.

Value

A data frame containing:

x

Input grid points.

y

Simulated noisy observations.

truef

True function values at x.

sd

Standard deviation used for noise at each point.

Examples

g <- function(x) sin(x / 5)
fashr:::simulate_data(g, sd = 0.2)
#>     x         y       truef  sd
#> 1   1 0.1896826  0.19866933 0.2
#> 2   2 0.3861803  0.38941834 0.2
#> 3   3 0.7534097  0.56464247 0.2
#> 4   4 0.8816003  0.71735609 0.2
#> 5   5 0.9602512  0.84147098 0.2
#> 6   6 1.1158346  0.93203909 0.2
#> 7   7 1.1418770  0.98544973 0.2
#> 8   8 1.0144866  0.99957360 0.2
#> 9   9 0.5759773  0.97384763 0.2
#> 10 10 1.0332626  0.90929743 0.2
#> 11 11 0.7972707  0.80849640 0.2
#> 12 12 0.6443041  0.67546318 0.2
#> 13 13 0.2213509  0.51550137 0.2
#> 14 14 0.2393581  0.33498815 0.2
#> 15 15 0.2247083  0.14112001 0.2
#> 16 16 0.2133618 -0.05837414 0.2