Perform Empirical Bayes (EB) on Inferred Likelihood Matrix

This function performs empirical Bayes estimation on the provided (log) likelihood matrix, incorporating an optional Dirichlet prior penalty.

fash_eb_est(L_matrix, penalty = 1, grid)

Arguments

L_matrix

A numeric matrix representing the log-likelihood values. Rows correspond to datasets, and columns correspond to grid points.

penalty

A numeric value representing the lambda value for the Dirichlet prior.

grid

A numeric vector representing the grid of PSD (Predictive Standard Deviation) values.

Value

A list containing:

prior_weight

A data frame with two columns: psd (non-trivial PSD values) and prior_weight (learned prior weights).

posterior_weight

A numeric matrix where rows correspond to datasets, and columns correspond to non-trivial PSD values, representing the posterior weights.

Examples

# Example usage
L_matrix <- matrix(rnorm(500), nrow = 100, ncol = 5)
grid <- seq(0, 2, length.out = 5)
result <- fash_eb_est(L_matrix, penalty = 2, grid = grid)
print(result$prior_weight)
#>   psd prior_weight
#> 1 0.0   0.31335429
#> 2 0.5   0.21984240
#> 3 1.0   0.17355377
#> 4 1.5   0.09710224
#> 5 2.0   0.19614730
print(result$posterior_weight)
#>                 0        0.5          1         1.5          2
#>   [1,] 0.74241670 0.04464474 0.07475874 0.046707002 0.09147281
#>   [2,] 0.37674187 0.13996955 0.16550654 0.282308580 0.03547346
#>   [3,] 0.19279313 0.57765277 0.05420684 0.114377213 0.06097005
#>   [4,] 0.82911896 0.04904574 0.06515758 0.042324858 0.01435286
#>   [5,] 0.44793740 0.14464583 0.06023914 0.039930656 0.30724698
#>   [6,] 0.29557004 0.14622190 0.09297339 0.338523109 0.12671156
#>   [7,] 0.76247501 0.07634919 0.04891426 0.013722903 0.09853863
#>   [8,] 0.14550764 0.27962631 0.04008105 0.060158436 0.47462656
#>   [9,] 0.40095200 0.05223151 0.06790996 0.038934851 0.43997168
#>  [10,] 0.19059498 0.06470527 0.52967372 0.180584213 0.03444181
#>  [11,] 0.29661401 0.14824975 0.26855532 0.043274856 0.24330606
#>  [12,] 0.17138085 0.04892482 0.09594302 0.085270104 0.59848121
#>  [13,] 0.41857728 0.19714756 0.16641249 0.110325786 0.10753689
#>  [14,] 0.41137574 0.20096965 0.03166842 0.134263296 0.22172290
#>  [15,] 0.10504764 0.19795305 0.27876871 0.172279576 0.24595102
#>  [16,] 0.18337235 0.36020021 0.33386824 0.020574545 0.10198465
#>  [17,] 0.09594723 0.07718491 0.07464673 0.064289728 0.68793141
#>  [18,] 0.15829053 0.19734819 0.46507841 0.050715398 0.12856746
#>  [19,] 0.28914647 0.21349560 0.22199402 0.026493864 0.24887005
#>  [20,] 0.50533777 0.02823346 0.28852046 0.004940814 0.17296750
#>  [21,] 0.15589314 0.47487216 0.06265777 0.155724638 0.15085230
#>  [22,] 0.17843092 0.20310926 0.33408626 0.057012436 0.22736113
#>  [23,] 0.55997341 0.04075160 0.26395683 0.008717866 0.12660030
#>  [24,] 0.10353860 0.10659855 0.37849509 0.240205354 0.17116241
#>  [25,] 0.29694870 0.21445117 0.08507059 0.037827239 0.36570230
#>  [26,] 0.14470356 0.41079608 0.15379229 0.123154082 0.16755399
#>  [27,] 0.11167160 0.65463761 0.02661524 0.146499387 0.06057615
#>  [28,] 0.25568544 0.11153945 0.08031218 0.171017569 0.38144536
#>  [29,] 0.42263625 0.08412162 0.21679374 0.124976588 0.15147180
#>  [30,] 0.42892795 0.19202532 0.18042608 0.077499985 0.12112067
#>  [31,] 0.06649305 0.38482195 0.04535932 0.415425773 0.08789990
#>  [32,] 0.30144639 0.40445205 0.17522770 0.018749326 0.10012454
#>  [33,] 0.05063275 0.42822950 0.25960274 0.068020425 0.19351459
#>  [34,] 0.75213822 0.16545477 0.01094030 0.037200642 0.03426605
#>  [35,] 0.17388712 0.29054025 0.27808750 0.129631802 0.12785334
#>  [36,] 0.18633748 0.49025026 0.24035295 0.065372572 0.01768674
#>  [37,] 0.32517245 0.06681420 0.04811571 0.177699303 0.38219834
#>  [38,] 0.42732054 0.10536241 0.26123341 0.004762396 0.20132123
#>  [39,] 0.42310866 0.06091953 0.09666843 0.178004431 0.24129895
#>  [40,] 0.53413124 0.21630974 0.14744015 0.041888067 0.06023081
#>  [41,] 0.13173348 0.59400959 0.16819816 0.072056820 0.03400195
#>  [42,] 0.30390823 0.05832803 0.31276727 0.189678429 0.13531804
#>  [43,] 0.50812671 0.23101377 0.02617159 0.043752639 0.19093529
#>  [44,] 0.19357525 0.31794493 0.29787463 0.086319439 0.10428575
#>  [45,] 0.50436251 0.02976773 0.36982359 0.070220585 0.02582559
#>  [46,] 0.21126247 0.38413532 0.14159488 0.135515085 0.12749224
#>  [47,] 0.13841112 0.20600240 0.09715817 0.035956764 0.52247155
#>  [48,] 0.06006654 0.22389908 0.18549133 0.250115339 0.28042770
#>  [49,] 0.30702540 0.27352379 0.16800002 0.161269618 0.09018116
#>  [50,] 0.02200959 0.12514491 0.29300744 0.048690340 0.51114772
#>  [51,] 0.06857957 0.42685493 0.10532238 0.088624487 0.31061864
#>  [52,] 0.61930801 0.03198689 0.19577877 0.041498482 0.11142784
#>  [53,] 0.21940684 0.51067925 0.04567499 0.060625526 0.16361340
#>  [54,] 0.65889519 0.06742027 0.05534942 0.048228626 0.17010649
#>  [55,] 0.30905439 0.18042570 0.20147278 0.090569935 0.21847719
#>  [56,] 0.36496541 0.07291413 0.35254430 0.065247096 0.14432907
#>  [57,] 0.11536839 0.02593719 0.41176272 0.059246540 0.38768515
#>  [58,] 0.18041476 0.74307094 0.05094691 0.004226224 0.02134117
#>  [59,] 0.30175566 0.12145560 0.37229276 0.094362872 0.11013311
#>  [60,] 0.48857584 0.15658468 0.06180506 0.018967997 0.27406642
#>  [61,] 0.09530551 0.36204262 0.26344064 0.075599500 0.20361173
#>  [62,] 0.28773732 0.35639323 0.14581495 0.119807446 0.09024705
#>  [63,] 0.14492073 0.09633952 0.24922641 0.302502211 0.20701114
#>  [64,] 0.19691119 0.35077484 0.10718307 0.030395225 0.31473568
#>  [65,] 0.24927078 0.28470920 0.01351193 0.039954644 0.41255345
#>  [66,] 0.14925046 0.11219005 0.29191090 0.067273143 0.37937544
#>  [67,] 0.29943520 0.23346597 0.37804663 0.029183385 0.05986881
#>  [68,] 0.48638669 0.18983906 0.21045710 0.014645402 0.09867175
#>  [69,] 0.30929745 0.55995783 0.07899040 0.022241958 0.02951236
#>  [70,] 0.22411960 0.44169418 0.06706862 0.051778188 0.21533941
#>  [71,] 0.13250913 0.24443910 0.28590115 0.253650343 0.08350027
#>  [72,] 0.09026538 0.38910948 0.02802397 0.028287965 0.46431321
#>  [73,] 0.23748007 0.09337646 0.08933420 0.501632410 0.07817685
#>  [74,] 0.51965556 0.20983723 0.07331563 0.149671948 0.04751963
#>  [75,] 0.62248175 0.01611871 0.06948453 0.081083479 0.21083153
#>  [76,] 0.08600301 0.02261700 0.04455678 0.088859317 0.75796388
#>  [77,] 0.18837434 0.12797916 0.33451009 0.321302785 0.02783362
#>  [78,] 0.51959643 0.12782257 0.14663831 0.042959066 0.16298363
#>  [79,] 0.13664030 0.18491672 0.09278584 0.121121701 0.46453543
#>  [80,] 0.10612166 0.11727450 0.07964392 0.039257617 0.65770230
#>  [81,] 0.11335232 0.64280703 0.10657274 0.062120142 0.07514777
#>  [82,] 0.32018872 0.45200858 0.08534708 0.046004702 0.09645092
#>  [83,] 0.12072114 0.15685286 0.46917980 0.086416851 0.16682934
#>  [84,] 0.27502145 0.24949087 0.23888419 0.068889583 0.16771392
#>  [85,] 0.12403086 0.48035403 0.08755008 0.028197012 0.27986802
#>  [86,] 0.21920346 0.15423700 0.16115985 0.048607574 0.41679212
#>  [87,] 0.04816565 0.14458261 0.62758662 0.098108764 0.08155636
#>  [88,] 0.60037627 0.07556370 0.04934648 0.040654960 0.23405859
#>  [89,] 0.70803060 0.06833988 0.11763451 0.026653722 0.07934129
#>  [90,] 0.04637567 0.22731802 0.41399681 0.022723246 0.28958625
#>  [91,] 0.76560722 0.09811215 0.04553726 0.070633076 0.02011030
#>  [92,] 0.65692218 0.16395003 0.03329102 0.046248025 0.09958875
#>  [93,] 0.80445772 0.01250964 0.02474774 0.076688165 0.08159673
#>  [94,] 0.07287595 0.04807165 0.60836937 0.063010070 0.20767296
#>  [95,] 0.28177982 0.17295655 0.29124554 0.172150381 0.08186771
#>  [96,] 0.67662307 0.15929836 0.01995048 0.039641697 0.10448639
#>  [97,] 0.40575680 0.06120421 0.01141647 0.363293667 0.15832885
#>  [98,] 0.29485058 0.37147217 0.08850791 0.190477475 0.05469187
#>  [99,] 0.24939290 0.10295568 0.26665318 0.040937095 0.34006114
#> [100,] 0.15820764 0.70914320 0.04896017 0.020100234 0.06358875